fix: local dev wiring and token-math correctness
- migrate.ts: skip re-running migrations whose tables were already seeded by docker-entrypoint-initdb.d
- pg-ai-generation.repository: cast integer token sums to numeric before multiplying by float rate constants to avoid integer arithmetic truncation
- gemini-client: log full 4xx error details (message + errorDetails body) to ease Gemini API debugging
- tsconfig.json: add typeRoots so local src/types declarations are resolved before @types
- ai_form_datasource.dart: hardcode LAN base URL for local device testing (was dart-define placeholder)
- google_auth_datasource.dart: pass serverClientId to GoogleSignIn for server-side token exchange
- subscription_service.dart: add kBypassPremium flag to skip RevenueCat during local testing

+++ gfm_mw/src/infrastructure/db/repositories/pg-ai-generation.repository.ts
@@ -158,8 +158,8 @@ export class PgAiGenerationRepository implements AiGenerationRepository {
   async getTotalSpendUsd(userId: number, sinceMs: number): Promise<number> {
     const { rows } = await this.db.query(
       `SELECT
-         COALESCE(SUM(input_tokens),  0) * $3
-       + COALESCE(SUM(output_tokens), 0) * $4 AS spend_usd
+         COALESCE(SUM(input_tokens),  0)::numeric * $3
+       + COALESCE(SUM(output_tokens), 0)::numeric * $4 AS spend_usd
        FROM ai_generations
        WHERE user_id    = $1
          AND created_at > to_timestamp($2 / 1000.0)`,
@@ -171,8 +171,8 @@ export class PgAiGenerationRepository implements AiGenerationRepository {
   async getGlobalDailySpendUsd(): Promise<number> {
     const { rows } = await this.db.query(
       `SELECT
-         COALESCE(SUM(input_tokens),  0) * $1
-       + COALESCE(SUM(output_tokens), 0) * $2 AS spend_usd
+         COALESCE(SUM(input_tokens),  0)::numeric * $1
+       + COALESCE(SUM(output_tokens), 0)::numeric * $2 AS spend_usd
        FROM ai_generations
        WHERE created_at >= date_trunc('day', NOW() AT TIME ZONE 'UTC')`,
       [INPUT_USD_PER_TOKEN, OUTPUT_USD_PER_TOKEN],
